Add notification when contributions received
This adds a message such as "🌑 Service node abc received a
contribution." plus some info about the current/required confirmations.
The moon can be one of: 🌑 for a minimal operator contribution (< 26%);
🌒 for a contribution up to 50%; 🌓 up to 75%, 🌔 above 75%, and 🌕 once
the final contribution makes 100%.  (This is sent in addition to the
"Service node is not fully staked and active message).